feat(proxy): add --proxy-domains flag for domain-specific proxying with wildcard support (#582)
- Rename “proxy-internal” to “proxy-target”.
- Rename “proxy-external” to “proxy-bright”. 
- Add a new “proxy-domains” flag as an optional flag. It can be accepted
when the user is setting either “proxy-target” or “proxy”. It can
receive a comma-separated list of domains which will be proxied. Any
domain that is not part of the list won’t be proxied. Notes: Support for
wildcard is required. If I specify “*.foo.bar” as one of the domains,
“a.foo.bar”, “b.foo.bar”, etc. will all be proxied.
- Bright domains will always be proxied when using with the “proxy”
flag, without a need to have them explicitly listed as part of the
“proxy-domains” list. if the user doesn’t want to proxy bright domains,
he should use only "proxy-target"