refactor(scan): replace deprecated attack parameter locations in RestScans configuration (#653)
There are 2 mistyped attack params:
* artifical-fragment
* artifical-query

But it's not accidental, because Scan API used to require exactly these
mistyped params.
To improve this, we amended our API to accept amended values. In this
change, we are
adding support to amended values:

```
  ARTIFICAL_FRAGMENT = 'artifical-fragment', // deprecated, use ARTIFICIAL_FRAGMENT instead
  ARTIFICAL_QUERY = 'artifical-query', // deprecated, use ARTIFICIAL_QUERY instead
  ARTIFICIAL_FRAGMENT = 'artificial-fragment',
  ARTIFICIAL_QUERY = 'artificial-query',
  ```

Then if users are using incorrect ones, replace them to amended ones that API requires.